Hi @gumyr thank you for this awesome library! 

this PR includes some small changes to the meshing logic that greatly improve the export speed for a couple of toy projects I'm working on (300K+ vertices), and wanted to share it as it may be helpful to others.

with change

| Number of vertices | Time (s) |
|-|-|
| 100 | 0.003 |
| 1000 | 0.002 |
| 10000 | 0.014 |
| 100000 | 0.215 |

without change

| Number of vertices | Time (s) |
|-|-|
| 100 | 0.004 |
| 1000 | 0.007 |
| 10000 | 0.509 |
| 100000 | 58.602 |


additional this PR adds a small test to ensure that `100K` vertices always take less than 1 second.

```bash
pytest tests/test_mesher.py -v
# tests/test_mesher.py::InternalApiBenchmark::test_create_3mf_mesh PASSED
```

Please let me know if I should make any changes, and thanks again! 🙏