@drbh 2025-01-14
just attempted to run `mypy` on both `dev` and this branch and they both seem to return the type issue so I'm not sure if these changes introduced the CI issue..