@gumyr 2025-01-14
Thanks!  300K vertices - that's a very large model. This change will be very helpful those that dream big.